스케줄러 활성화, 금액관련 변수타입 변경, 로깅 포맷변경

+++ src/main/java/my/trader/coin/model/Trade.java
@@ -4,6 +4,8 @@ import lombok.Getter;
 import lombok.Setter;
 
 import jakarta.persistence.*;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -17,8 +19,8 @@ public class Trade {
     private Long id;
 
     private String type; // BUY or SELL
-    private double price;
-    private double quantity;
+    private BigDecimal price;
+    private Double quantity;
     private LocalDateTime timestamp;
 
     // getters and setters
